// Utils/interface.ts
// export default interface Books {
//     title: string;
//     description: string;
//     author: string;
//     views: []
//     category: string
// }

// export default interface Books{
//     title: string;
//     category: string;
//     author: string;
//     coverImage: string;
//     views: [];
// }

// export default interface Books{
//     title: string;
//     description: string;
//     author: string;
//     category: string
//     views: [];   
// }

export default interface Books {
    title: string;
    description: string;
    author: string;
    category: string
    coverImage: string;
    views: [];
}